requirements.txt
flask==2.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
flask-cors==4.0.0 